Major enhancement:
1. Usermanagement module added.
 - Useradd | delete | passwd | enable | disable | list
2. Fixed the bug when typing passwords
3. Optimized the init process.
4. Excluded Desktop environment installation for compute
    nodes.

+++ scripts/hpcmgr.sh
@@ -11,7 +11,7 @@ if [ $CURRENT_USER != 'root' ]; then
   echo -e "[ FATAL: ] *ONLY* root user can run the command 'hpcmgr'. "
   echo -e "           Please make sure you use either user1 with 'sudo' privilege, OR"
   echo -e "           use root (NOT recommend!) to run 'hpcmgr'. Exit now."
-  exit
+  exit 1
 fi
 #CRITICAL: Environment Variable $NODE_NUM and $NODE_CORES MUST be written to /etc/profile IN ADVANCE!
 source /etc/profile
@@ -45,64 +45,65 @@ function node_invalid_info() {
 
 if [ ! -n "$1" ]; then
   help_info
-  exit
+  exit 3
 fi
 
 if [ $1 != 'quick' ] && [ $1 != 'master' ] && [ $1 != 'all' ] && [ $1 != 'clear' ] && [ $1 != 'users' ] && [ $1 != 'connect' ] && [ $1 != 'install' ]; then
   help_info
-  exit
+  exit 3
 fi
 
 if [ $1 = 'clear' ]; then
   if [ ! -f /root/hostfile ]; then
     node_invalid_info
-    exit
+    exit 5
   fi
   rm -rf /root/hostfile_dead_nodes && touch /root/hostfile_dead_nodes
   echo -e "[ -INFO- ] The hostfile_dead_nodes list had been emptied. Exit now."
   echo -e "[ -INFO- ] The hostfile_dead_nodes list had been emptied." >> ${logfile}
-  exit
+  exit 0
 fi
 
 ##### USER MANAGEMENT #####################
 if [ $1 = 'users' ]; then
   if [ ! -f /root/hostfile ]; then
     node_invalid_info
-    exit
+    exit 5
   fi
   if [ ! -n "$2" ]; then
     echo -e "Usage: \n\thpcmgr users list\n\thpcmgr users add your_user_name your_password\n\thpcmgr users delete your_user_name\nPlease check your parameters. Exit now.\n"
-    exit
+    exit 3
   fi
   if [[ $2 != 'list' && $2 != 'add' && $2 != 'delete' ]]; then
     echo -e "Usage: \n\thpcmgr users list\n\thpcmgr users add your_user_name your_password\n\thpcmgr users delete your_user_name\nPlease check your parameters. Exit now.\n"
-    exit
+    exit 3
   fi
   if [ $2 = 'list' ]; then
     sacctmgr list user
-    exit
+    exit 0
   fi
   if [ $2 = 'add' ]; then
     if [ ! -n "$3" ]; then
       echo -e "Usage: \n\thpcmgr users add your_user_name your_password\nPlease check your parameters. Exit now.\n"
-      exit
+      exit 3
     else
       id $3
       if [ $? -eq 0 ]; then
         sacctmgr list user | grep -w $3
         if [ $? -eq 0 ]; then
           echo -e "User $3 already exists in this cluster. Please specify another username. Exit now. \n"
-          exit
+          exit 11
         else
           echo -e "User $3 already exists in this OS but not in this cluster. Adding to the cluster now.\n"
           echo "y" | sacctmgr add user $3 account=hpc_users
-          exit
+          sed -i "/$3/,+0 s/DISABLED/ENABLED/g" $user_registry
+          exit 0
         fi
       else
         useradd $3 -m
         if [ -n "$4" ]; then
           echo "$4" | passwd $3 --stdin > /dev/null 2>&1
-          echo -e "username: $3 $4" >> $user_registry
+          echo -e "username: $3 $4 ENABLED" >> $user_registry
         else
           echo -e "Generate random string for password." 
           if [ $CENTOS_V -eq 7 ]; then
@@ -112,7 +113,8 @@ if [ $1 = 'users' ]; then
           fi
           cat /root/.cluster_secrets/secret_$3.txt | passwd $3 --stdin > /dev/null 2>&1
           echo -n "username: $3 " >> $user_registry
-          cat /root/.cluster_secrets/secret_$3.txt >> $user_registry
+          new_passwd=`cat /root/.cluster_secrets/secret_$3.txt`
+          echo -e "$new_passwd ENABLED" >> $user_registry
           rm -rf /root/.cluster_secrets/secret_$3.txt
         fi
         mkdir -p /home/$3/.ssh
@@ -127,14 +129,17 @@ if [ $1 = 'users' ]; then
         ln -s /hpc_apps /home/$3/Desktop/ >> /dev/null 2>&1
         for i in $(seq 1 $NODE_NUM )
         do
-          ssh -n compute${i} "useradd $3 -m"
-          user_passwd=`cat $user_registry | grep $3 | awk '{print $3}'`
-          ssh -n compute${i} "echo '$user_passwd' | passwd $3 --stdin > /dev/null 2>&1"
-          ssh -n compute${i} "rm -rf /home/$3/.ssh"
-          ssh -n compute${i} "mkdir -p /home/$3/Desktop && ln -s /hpc_apps /home/$3/Desktop/ && ln -s /hpc_data/$3_data /home/$3/Desktop/"
-          scp -r -q /home/$3/.ssh root@compute${i}:/home/$3/
-          scp -q /home/$3/Desktop/*desktop root@compute${i}:/home/$3/Desktop/
-          ssh -n compute${i} "chown -R $3:$3 /home/$3"
+          ping -c 2 -W 1 -q compute${i} >> ${logfile} 2>&1
+          if [ $? -eq 0 ]; then
+            ssh -n compute${i} "useradd $3 -m"
+            #user_passwd=`cat $user_registry | grep $3 | awk '{print $3}'`
+            #ssh -n compute${i} "echo '$user_passwd' | passwd $3 --stdin > /dev/null 2>&1"
+            ssh -n compute${i} "rm -rf /home/$3/.ssh"
+            #ssh -n compute${i} "mkdir -p /home/$3/Desktop && ln -s /hpc_apps /home/$3/Desktop/ && ln -s /hpc_data/$3_data /home/$3/Desktop/"
+            scp -r -q /home/$3/.ssh root@compute${i}:/home/$3/
+            #scp -q /home/$3/Desktop/*desktop root@compute${i}:/home/$3/Desktop/
+            ssh -n compute${i} "chown -R $3:$3 /home/$3"
+          fi
         done
         if [ -f /root/.cos.conf ]; then
           cp /root/.cos.conf /home/$3/ && chown -R $3:$3 /home/$3/.cos.conf
@@ -146,24 +151,24 @@ if [ $1 = 'users' ]; then
           cp /root/.s3cfg /home/$3/ && chown -R $3:$3 /home/$3/.s3cfg
         fi
         echo "y" | sacctmgr add user $3 account=hpc_users
-        exit
+        exit 0
       fi
     fi
   fi
   if [ $2 = 'delete' ]; then
     if [ ! -n "$3" ]; then
-      echo -e "Usage: \n\thpcmgr users delete your_user_name\n\thpcmgr users delete your_user_name os\nPlease check your parameters. Exit now.\n"
-      exit
+      echo -e "Usage: \n\thpcmgr users delete your_user_name\n\thpcmgr users delete your_user_name os\nPlease check your parameters. Exit now."
+      exit 3
     fi
-    if [[ -n "$3" && $3 = 'root' ]]; then
-      echo -e "ROOT USER cannot be deleted! Exit now. \n"
-      exit
+    if [[ $3 = 'root' || $3 = 'user1' ]]; then
+      echo -e "ROOT USER and User1 cannot be deleted! Exit now."
+      exit 13
     fi
     getname=`sacctmgr list user | grep -w $3 | awk '{print $1}'`
     if [ ! -n "$4" ] || [ $4 != 'os' ]; then
       if [ "$getname" != "$3" ]; then
         echo -e "$3 is not in the cluster. Nothing deleted, exit now."
-        exit
+        exit 15
       fi
     fi
     if [ "$getname" = "$3" ]; then
@@ -171,17 +176,21 @@ if [ $1 = 'users' ]; then
     fi
     if [[ ! -n "$4" || $4 != "os" ]]; then
       echo -e "User $3 has been deleted from the cluster, but still in the OS."
-      exit
+      sed -i "/$3/,+0 s/ENABLED/DISABLED/g" $user_registry
+      exit 0
     else
       echo -e "[ -WARN- ] User $3 will be erased from the Operating System permenantly!"
       userdel -f -r $3 && mv /hpc_data/${3}_data /hpc_data/${3}_data_deleted_user
       for i in $(seq 1 $NODE_NUM )
       do
-        ssh -n compute${i} "userdel -f -r $3"
+        ping -c 2 -W 1 -q compute${i} >> ${logfile} 2>&1
+        if [ $? -eq 0 ]; then
+          ssh -n compute${i} "userdel -f -r $3"
+        fi
       done
       sed -i "/$3/d" $user_registry
       echo -e "[ -DONE- ] User $3 Deleted permenantly."
-      exit
+      exit 0
     fi 
   fi
 fi
@@ -190,7 +199,7 @@ fi
 if [ $1 = 'quick' ]; then
   if [ ! -f /root/hostfile ]; then
     node_invalid_info
-    exit
+    exit 3
   fi
   sacct >> /dev/null 2>&1
   if [ $? -eq 0 ]; then
@@ -198,7 +207,7 @@ if [ $1 = 'quick' ]; then
     if [ $? -eq 0 ]; then
       echo -e "[ -WARN- ] There are still running tasks in current cluster. You need to cancel all the tasks first."
       echo -e "[ -WARN- ] Exit now.\n"
-      exit
+      exit 17
     fi
   fi  
   echo -e "[ -INFO- ] Welcome to the quick mode. All services will be restarted now."
@@ -226,14 +235,14 @@ if [ $1 = 'quick' ]; then
   done
   echo -e "[ -DONE- ] HPC-NOW Cluster Status:\n"
   sinfo -N
-  exit
+  exit 0
 fi
 
 ####### Check connectivities ###########################
 if [ $1 = 'connect' ]; then
   if [ ! -f /root/hostfile ]; then
     node_invalid_info
-    exit
+    exit 3
   fi
   compute_passwd=`cat /root/.cluster_secrets/compute_passwd.txt`
   echo -e "[ STEP 1 ] Checking the network connection now ... "
@@ -336,7 +345,7 @@ if [ $1 = 'connect' ]; then
   else
     echo -e "[ FATAL: ] PLEASE MAKE SURE the /etc/hosts-clean exists.\n Exit now."
     echo -e "[ FATAL: ] PLEASE MAKE SURE the /etc/hosts-clean exists.\n Exit now." >> ${logfile}
-    exit
+    exit 31
   fi
   echo -e "[ STEP 5 ] Setting up users of compute nodes ... "
   for i in $(seq 1 $NODE_NUM )
@@ -345,32 +354,32 @@ if [ $1 = 'connect' ]; then
     while read hpc_user_row
     do
       username=`echo -e $hpc_user_row | awk '{print $2}'`
-      user_passwd=`echo -e $hpc_user_row | awk '{print $3}'`
-      ssh -n compute${i} "echo '$user_passwd' | passwd $username --stdin >> /dev/null 2>&1"
+      #user_passwd=`echo -e $hpc_user_row | awk '{print $3}'`
+      #ssh -n compute${i} "echo '$user_passwd' | passwd $username --stdin >> /dev/null 2>&1"
       scp -r -q /home/$username/.ssh root@compute${i}:/home/$username/
       ssh -n compute${i} "chown -R $username:$username /home/$username >> /dev/null 2>&1"
     done < $user_registry
   done
   echo -e "[ STEP 5 ] Users are ready."
   echo -e "[ -DONE- ] Connectivety check finished! \n[ -DONE- ] You need to run the command 'hpcmgr all' on the master node.\n[ -DONE- ] Exit now."
-  exit
+  exit 0
 fi
 # all mode: restart services on all nodes
 if [[ $1 = 'master' || $1 = 'all' ]]; then
   if [ ! -f /root/hostfile ]; then
     node_invalid_info
-    exit
+    exit 3
   fi
   if [ $((NODE_NUM)) -lt $((1)) ]; then
     echo -e "[ FATAL: ] There is no compute nodes in the current cluster. You need to add at least 1 compute node to start the cluster.\n[ FATAL: ] Exit now.\n"
-    exit
+    exit 21
   fi 
   sacct >> /dev/null 2>&1
   if [ $? -eq 0 ]; then
     sacct | grep running  >> /dev/null 2>&1
     if [ $? -eq 0 ]; then
       echo -e "[ -WARN- ] There are still running tasks in current cluster. You need to cancel all the tasks first.\n[ -WARN- ] Exit now.\n"
-      exit
+      exit 17
     fi
   fi
 # To make sure the munged is up
@@ -383,7 +392,7 @@ if [[ $1 = 'master' || $1 = 'all' ]]; then
     if [ $? -ne 0 ]; then
       echo -e "[ FATAL: ] FAILED to start munge daemon on Master Node! Please check the installation of munge.\nExit now."
       echo -e "[ FATAL: ] FAILED to start munge daemon on Master Node! Please check the installation of munge.\nExit now." >> ${logfile}
-      exit
+      exit 33
     fi
   fi
   echo -e "[ STEP 1 ] Authentication of the master node is OK."
@@ -396,7 +405,7 @@ if [[ $1 = 'master' || $1 = 'all' ]]; then
     if [ $? -ne 0 ]; then
       echo -e "[ FATAL: ] FAILED to start munge daemon on Compute${i} Node! Please check the installation of munge.\nExit now."
       echo -e "[ FATAL: ] FAILED to start munge daemon on Compute${i} Node! Please check the installation of munge.\nExit now." >> ${logfile}
-      exit
+      exit 33
     fi
   done
   echo -e "[ STEP 2 ] Authentication of Compute node(s) is OK."
@@ -413,13 +422,13 @@ if [[ $1 = 'master' || $1 = 'all' ]]; then
   else
     echo -e "[ FATAL: ] SLURMDBD is not properly started. Exit now."
     echo -e "[ FATAL: ] SLURMDBD is not properly started.\n" >> ${logfile}
-    exit  
+    exit 33
   fi
   systemctl status slurmctld | grep "Active: failed" >> ${logfile}
   if [ $? -eq 0 ]; then
     echo -e "[ FATAL: ] SLURMCTLD is not properly started. Exit now."
     echo -e "[ FATAL: ] SLURMCTLD is not properly started.\n" >> ${logfile}
-    exit
+    exit 33
   fi
    
   if [ $1 = 'all' ]; then
@@ -461,7 +470,7 @@ fi
 if [ $1 = 'install' ]; then
   if [ -z $URL_INSTSCRIPTS_ROOT ]; then
     echo -e "[ FATAL: ] Failed to connect to a valid appstore repo. Exit now."
-    exit
+    exit 35
   fi
   if [[ ! -n "$2" || $2 = 'list' ]]; then
     echo -e "[ -INFO- ] Usage: hpcmgr install software_to_be_installed"
@@ -505,7 +514,7 @@ if [ $1 = 'install' ]; then
     echo -e "\tkswps     - WPS Office Suite for Linux (RECOMMENDED)" 
     echo -e "\tenvmod    - Environment Modules" 
     echo -e "\tvscode    - Visual Studio Code" 
-    exit
+    exit 0
   elif [[ -n $2 && $2 = 'of7' ]]; then
     curl -s ${URL_INSTSCRIPTS_ROOT}install_of7.sh | bash
   elif [[ -n $2 && $2 = 'of2112' ]]; then
